backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
click==8.3.1
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
jose==1.0.0
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==1.0.0
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
